chore: Extracts function to utilities and adds warn message
